femenuoptions TransitionOn__17tUserNameMenuItem: 219 -> 6, insn-EXACT 141/141
SYM-first rewrite: the 8c block lists exactly two locals, `short i` (REG $s2)
and `short NumberOfRows[6]` (AUTO -0x28); the six other Ghidra temps and the
hand-expanded unaligned-store soup were fictions.  NumberOfRows is an AGGREGATE
INITIALIZER `{7,9,9,9,8,9}` (bytes read off asm/data D_80010A00) = gcc's own
12-byte rodata->stack block copy (3x lwl/lwr + swl/swr).  Row/column scan
re-derived from the oracle + SLD 1982-1996 as two rotated
`while (c != '!' && c != '@' && <bound>)` loops.

Three declaration-shape fixes on top:
 * CORRECTNESS: `menu_kUserNameRows` was declared `extern int` but the SYM says
   `00529b0 94 Def class EXT type SHORT` -- the oracle stores it with `sh` and
   reads it with `lh`.  (screenusername.cpp re-gated: byte-identical, no change.)
 * the scalar extern is a DELAY-SLOT POISON PILL (catalog): cc1 emits `lh $r,sym`
   / `sh $r,sym` as one assembler MACRO -- unschedulable, and the store form
   burns `$at`.  An unsized-array asm-label view (`extern short X[] __asm__(X)`)
   splits it into a schedulable lui + lh/sh pair, which is also what fills the
   entry guard's branch delay slot.
 * same lever on FEApp gives the oracle's `lui $s2,%hi(FEApp)` held across the
   FeAudio_AsyncPlaySpeech call (a plain `T **p = &FEApp;` local does NOT --
   gcc folds it straight back).

Residual 6 = one 3-instruction sched1 permutation (the two %hi materializations
vs the frontEnd.language lbu), at exact instruction parity.

+++ recon/frontend/common/femenuoptions_externs.h
@@ -31,7 +31,7 @@ void  AudioCmn_PlayFESFX(int);  void AudioCmn_PlayFESFXVol(int,int);
 /* batch externs */
 extern tScreenAudio *screenAudio;  extern tScreenControllerConfig *screenControllerConfig[];   /* unsized array (§3.12 #5): oracle lui/lw separate-scratch */
 extern tTexture_ShapeInfo *gHelpShapes;  extern tPlayer gMenu_SubMenuPlayer;
-extern int menu_kUserNameRows;
+extern short menu_kUserNameRows;   /* SYM: $800529b0 94 Def class EXT type SHORT (was wrongly int -> sw/lw instead of sh/lh) */
 int CalcFadeVal(int,int,int);  int CalcFadeVal(int,int,int,int);
 void FETextRender_FullTextRGB(char*,short,short,int,char,short);
 void FETextRender_FullText(char*,short,short,tMenuTextType,tMenuTextState,short);
